requirements.txt
streamlit
cohere
youtube-transcript-api
langchain
python-dotenv
